Create REBINDG
This query will read SYSIBM.SYSPACKAGE and generate BIND PACKAGE statements for everything it finds. You can add filter predicates to limit the resultset to the end of it.
The ORDER BY CONTOKEN clause is vital here because it will stop duplicate lines being created which sometimes happens if you have a package with a lot of duplicate information (collid, name etc).